+++ pkg/helm/testdata/empty-repo-index.yaml
@@ -1,0 +1,1 @@
+entries:
