+++ pkg/helm/testdata/valid-index.yaml
@@ -1,0 +1,69 @@
+entries:
+  acs-engine-autoscaler:
+    - apiVersion: v1
+      appVersion: 2.1.1
+      created: 2017-12-06T18:48:59.568323124Z
+      description: Scales worker nodes within agent pools
+      digest: 39e66eb53c310529bd9dd19776f8ba662e063a4ebd51fc5ec9f2267e2e073e3e
+      icon: https://github.com/kubernetes/kubernetes/blob/master/logo/logo.png
+      maintainers:
+        - email: ritazh@microsoft.com
+          name: ritazh
+        - email: wibuch@microsoft.com
+          name: wbuchwalter
+      name: acs-engine-autoscaler
+      sources:
+        - https://github.com/wbuchwalter/Kubernetes-acs-engine-autoscaler
+      urls:
+        - https://kubernetes-charts.storage.googleapis.com/acs-engine-autoscaler-2.1.1.tgz
+      version: 2.1.1
+  wordpress:
+    - appVersion: 4.9.1
+      created: 2017-12-06T18:48:59.644981487Z
+      description: new description!
+      digest: 74889e60a35dcffa4686f88bb23de863fed2b6e63a69b1f4858dde37c301885c
+      engine: gotpl
+      home: http://www.wordpress.com/
+      icon: https://bitnami.com/assets/stacks/wordpress/img/wordpress-stack-220x234.png
+      keywords:
+        - wordpress
+        - cms
+        - blog
+        - http
+        - web
+        - application
+        - php
+      maintainers:
+        - email: containers@bitnami.com
+          name: bitnami-bot
+      name: wordpress
+      sources:
+        - https://github.com/bitnami/bitnami-docker-wordpress
+      urls:
+        - https://kubernetes-charts.storage.googleapis.com/wordpress-0.7.5.tgz
+      version: 0.7.5
+    - appVersion: 4.9.0
+      created: 2017-12-01T11:49:00.136950565Z
+      description: Web publishing platform for building blogs and websites.
+      digest: a69139ef3008eeb11ca60261ec2ded61e84ce7db32bb3626056e84bcff7ec270
+      engine: gotpl
+      home: http://www.wordpress.com/
+      icon: https://bitnami.com/assets/stacks/wordpress/img/wordpress-stack-220x234.png
+      keywords:
+        - wordpress
+        - cms
+        - cms
+        - blog
+        - http
+        - web
+        - application
+        - php
+      maintainers:
+        - email: containers@bitnami.com
+          name: bitnami-bot
+      name: wordpress
+      sources:
+        - https://github.com/bitnami/bitnami-docker-wordpress
+      urls:
+        - https://kubernetes-charts.storage.googleapis.com/wordpress-0.7.4.tgz
+      version: 0.7.4
