+++ pkg/http-client/httpclient.go
@@ -1,0 +1,52 @@
+/*
+Copyright © 2021 VMware
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package httpclient
+
+import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+)
+
+type Client interface {
+	Do(req *http.Request) (*http.Response, error)
+}
+
+func Get(url string, cli Client, headers map[string]string) ([]byte, error) {
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	for header, content := range headers {
+		req.Header.Set(header, content)
+	}
+
+	res, err := cli.Do(req)
+	if res != nil {
+		defer res.Body.Close()
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	if res.StatusCode != http.StatusOK {
+		errC, err := ioutil.ReadAll(res.Body)
+		if err != nil {
+			return nil, fmt.Errorf("failed to read error content: %v", err)
+		}
+		return nil, fmt.Errorf("request failed: %v", string(errC))
+	}
+
+	return ioutil.ReadAll(res.Body)
+}
